organize(haskell): move more code to HostFunction.hs
